+++ src/Helpers/FilterHelper.php
@@ -45,13 +45,45 @@ class FilterHelper
             } else {
                 $searchBodyFilters[] = [
                     [
-                        Filters::NAME  => $newPrefix,
-                        Filters::VALUE => $value,
+                        Filters::NAME => $newPrefix,
+                        Filters::VALUE => is_array($value) ? $this->sanitizeArray($value) : $this->sanitizeValue($value),
                     ],
                 ];
             }
         }
 
         return array_merge(...$searchBodyFilters);
+    }
+
+    /**
+     * Makes sure that the value is a string.
+     *
+     * If the original value is a bool: "true" or "false" as a string
+     * Otherwise, just the original value as a string
+     *
+     * @param $value
+     * @return string
+     */
+    private function sanitizeValue($value): string
+    {
+        if (is_bool($value)) {
+            return $value ? 'true' : 'false';
+        }
+
+        return (string)$value;
+    }
+
+    /**
+     * Runs the sanitizeValue method on each of the values of the array.
+     *
+     * @param array $values
+     * @return array
+     */
+    private function sanitizeArray(array $values): array
+    {
+        return array_map([
+            $this,
+            'sanitizeValue',
+        ], $values);
     }
 }